Retire real_train + quick_train + the arq worker
#265 Phase 2. The frontend stopped offering the runtime-adaptation
modes in #266; this PR removes the backend half so the API surface,
the schemas, the worker container, and the redis-backed job queue all
go with it. Encoder fine-tuning still happens, just offline through
the dedicated training scripts -- not on the live request path.

- app/main.py: arq pool lifespan, the _train_jobs dict, the
  _redis_pool helper, _set_job_state, _run_real_train_job + _body,
  _enqueue_real_train, the /train-jobs detail + listing endpoints,
  and every Redis state-marshalling helper come out. The /analyze
  handler collapses to a single fast path with a cold-start bootstrap
  branch; COLD_START_HISTORY_LENGTH (252d) replaces the worker
  module's REAL_TRAIN_HISTORY_LENGTH.
- app/schemas.py: AnalyzeRequest loses forecast_mode;
  TrainJobAcceptedResponse, TrainJobStatusResponse, TrainJobSummary,
  TrainJobsListResponse are removed.
- app/services/forecaster.py: the quick_train adaptation branch in
  forecast_quantitative_series is replaced with a single _get_model()
  call; the forecast_mode parameter stays for back-compat with
  persisted history rows but is no longer load-bearing.
- app/worker.py and tests/unit/test_arq_worker.py and
  tests/unit/test_real_train_endpoints.py: deleted.
- docker-compose.yml: worker, worker-gpu, and redis services
  removed; REDIS_URL env + depends_on: redis drops from both backend
  services.
- pyproject.toml: arq, redis, and fakeredis dependencies dropped.
- tests/contract/test_diff_modes.py renamed to test_analyze_contract
  and collapsed to a single fast-mode contract.
- Other tests touched to drop forecast_mode from /analyze payloads or
  to assert the new strict-rejection behaviour.

Tracks #265.
