+++ backend/pyproject.toml
@@ -40,8 +40,6 @@ dependencies = [
   "httpx>=0.27.0",
   "SQLAlchemy>=2.0",
   "structlog>=24.1",
-  "arq>=0.26",
-  "redis>=5.0",
 ]
 
 [project.optional-dependencies]
@@ -55,7 +53,6 @@ dev = [
   "py-spy>=0.3.14",
   "types-PyYAML",
   "types-requests",
-  "fakeredis>=2.20",
   "schemathesis>=3.30",
 ]
 gpu = []
