+++ tests/unit/test_main_api.py
@@ -102,7 +102,6 @@ def test_analyze_happy_path_with_realized_overlay(monkeypatch):
             "text": "sample",
             "date": "2026-03-15",
             "symbol": "^GSPC",
-            "forecast_mode": "fast",
             "horizon": "3d",
             "include_realized": True,
         },
@@ -120,7 +119,11 @@ def test_analyze_happy_path_with_realized_overlay(monkeypatch):
     assert payload["model"]["chunk_attention"]["lambda_value"] == pytest.approx(0.1234)
 
 
-def test_analyze_invalid_mode_returns_422():
+def test_analyze_rejects_unknown_field():
+    """Strict request schema (#265 Phase 2) — forecast_mode is no longer
+    a field, so a stale payload that includes it must 422 rather than
+    silently land on the fast-mode path."""
+
     client = TestClient(main_mod.app)
     response = client.post(
         "/analyze",
@@ -128,7 +131,7 @@ def test_analyze_invalid_mode_returns_422():
             "text": "sample",
             "date": "2026-03-15",
             "symbol": "^GSPC",
-            "forecast_mode": "bad_mode",
+            "forecast_mode": "fast",
             "horizon": "3d",
         },
     )
